grub.cfg-recovery: if mode + system are from env, only use those
If the chooser set both the mode and the system on our grubenv, then we
should obey that strictly and not look for other systems, as the chooser
is expecting us to boot/use that combo.

However, if we don't have both the mode + system set, then we need to
load everything and "pick" the best one.

Signed-off-by: Ian Johnson <person.uwsome@gmail.com>